feat: Allow to work with real sizes instead of pixel values (#194)
* chore: First ruler try

* chore: Allow to select pixel size

* chore: Finished size settings

* chore: Fixed many warnings

* chore: Some more warning fixes

* chore: Further warning fix

* chore: Further fixes

* chore: Further warning fixes

* chore: Removed all warnings

* chore: Added real size

* chore: Unit suffix change

* chore: Add jump box

* chore: Jump spinner

* chore: First working pixel val

* chore: LAst fix in unit conversion

* chore: Further watning fix

* chore: Fixed load pixel size

* chore: No position in overview

* chore: Select group by name

* chore: Ruker fix

+++ src/backend/commands/classification/ai_classifier/models/yolo/ai_model_yolo.hpp
@@ -24,8 +24,8 @@ class AiModelYolo : public AiModel
 public:
   struct ProbabilitySettings
   {
-    float maskThreshold  = 0.8;
-    float classThreshold = 0.5;
+    float maskThreshold  = 0.8F;
+    float classThreshold = 0.5F;
   };
 
   AiModelYolo(const ProbabilitySettings &settings);
@@ -33,18 +33,18 @@ public:
 
 private:
   /////////////////////////////////////////////////////
-  static constexpr inline int SEG_WIDTH       = 160;
-  static constexpr inline int SEG_HEIGHT      = 160;
-  static constexpr inline int NET_WIDTH       = 640;
-  static constexpr inline int NET_HEIGHT      = 640;
-  static constexpr inline float BOX_THRESHOLD = 0.25;    // (default = 0.25)
-  static constexpr inline float NET_STRIDE[4] = {8, 16, 32, 64};
-  static constexpr inline int SEG_CHANNELS    = 32;
-  static constexpr inline int STRIDE_SIZE     = 3;
-  static constexpr inline float NMS_THRESHOLD = 0.45;    // To prevent double bounding boxes (default = 0.45)
+  static constexpr inline int SEG_WIDTH        = 160;
+  static constexpr inline int SEG_HEIGHT       = 160;
+  static constexpr inline int NET_WIDTH        = 640;
+  static constexpr inline int NET_HEIGHT       = 640;
+  static constexpr inline float BOX_THRESHOLD  = 0.25;    // (default = 0.25)
+  static constexpr inline float NET_STRIDE[4]  = {8, 16, 32, 64};
+  static constexpr inline int64_t SEG_CHANNELS = 32;
+  static constexpr inline int STRIDE_SIZE      = 3;
+  static constexpr inline float NMS_THRESHOLD  = 0.45F;    // To prevent double bounding boxes (default = 0.45)
 
   const float mClassThreshold = 0.5;
-  const float mMaskThreshold  = 0.8;
+  const float mMaskThreshold  = 0.8F;
   const float mNmsScoreThreshold;
 
   /////////////////////////////////////////////////////
